Add files via upload
Modiefied ref_gl.dll files for "Soldier of Fortune 1". Commands that exist in game have been modified to be used in multiplayer, "r_fullbright - r_dullbright 1 and 0 / on and off"
"gl_drawmode- gl_grawmode 1 and 0 / on and off" and "gl_drawflat-gl_grawflat 1 and 0 / on and off. Just isolate existing ref_gl.dll file and copy and paste these in sof main folder.